+++ src/main/java/com/example/ecom/Entity/ProductEntity.java
@@ -29,5 +29,5 @@ public class ProductEntity {
     private Double price;
 
     @Column(name = "category_id",columnDefinition = "int DEFAULT NULL")
-    private int category;
+    private Long category;
 }
